Add new file-container for package-managers
File package-managers.ts was added.
